Chore: Busy making the cahnge from Time to Date.
The change from Time to Date came as a requirement of the pause feature. Bear in mind Time only takes into account second and minute while Date starts from millisecond all the way to year. As such let imagine a situation where (we still keep the Time object) the game is paused for a year (I am not sure how that is even possible) or realistically for a month (this seems doable). With the Time, everything gets srewed up in terms of timing (e.g. events do not fire, the clock in the top panel goes haywire etc...). On the other hand should we be timing according to date, all that is required is sace the date before pausing, and once we are back from the pause, we get the current date, perform the date difference and add it to all the existing dates. More importantly the clock's one. And we are golden.