Chore: Fixed the timing inconsistency of the clock, when coming back from pause.
Before this fix, what was happening is, coming back from pause the clock
in the top panel does not take into account the pause duration. In order
word this clock "keeps" running even wehn the game is paused. In fact
the clock should stop when the game is paused, and only restart when the
game is restarted again. After the fix, this now taking place. The way I
fixed the issue is by recording the date before going into pause, and
add the pause duration to the the play duration in the clock object once
we are back from pause.