chore: Just fixed the column without cubes issue.
Initially what was happening is, when generating a new
column starting date, I keep looping until I have a
date that is at most 0.5 second into the future. In fact
what I wanted to do was generate a date that is at least
0.5 seconds into the future. I fixed that, played a number
of games and did not encountered the issue. However I
could be wrong and the issue would still persist.