chore: Fixed the no cube in column issue.
I think I definitly have fixed the issue. The way a new event was being
added to the event queue before was incomplete. I was looping over the
internal queue of the event queue until I find an event with sheculed
time greater than the new incoming event. But the question is the
following, what if the new incoming event has the latest time? I did not
handle that part. I do now by checking whether the new event has been
added to the queue at the end the for loop. If not, then I just push it
at the end of the queue.