feat: Add background image.
The way it is now, it is resommended to run the application from a
simple web server, and doing so, the background of the canvas would
be set to a gray scale image of the Milky Way. However, even if the
user fail to do so and cross origin request not be allowed on the
browser, instead of game crashing, the background would be set to
a dark gray colour.