chore: fixed emotion issue in side panel
Refering to the previous commit, the issue was due to the sequential
process. What was happening is, the emoticon gets reset to normal face
(through reset of the side panel) in the the endGame function. Thereafter,
it gets set back to smiley at the end of the cube collision handler. So
to fix it, I created an end play event that get scheduled once the end
of play returns true inside player. This event has priority over all
other and automatically gets pushed to the from of the queue. Now
processing finishes and then endOfPlay gets called.