chore: chnage demarcation height for cube and add game exception
The demarcation height is the height at which a pn cube is allowed to schedule a new start event. Initially it was set to 1/3 of the height of the canvas. Now change it to somewhere between 20% - 45% of the canvas height.

Also add a game exception object that can be throw whenever something shaddy takes place.