fix: No issue found
There was no issue found with applying penalty for collecting an already collected (which is not yet to be collected) pn cube. There was no test case for the scenario of applying penalty for collecting an already collected (which is not yet to be collected) pn cube. It has beeen now added and all tests are passing.