Fix: Sort issue with new game button
What was happening before was the following, the new game button does not actually restart the game. This was because a call recreate the pn cube comes back with an error. That specifc function being invoked was initially global, but later on I moved it to utilitites. That was the cause for the issue.